🚚 Move MarkdownDocument to own sub domain
