+++ tests/RouteResolverTest.php
@@ -4,7 +4,7 @@ namespace Aheenam\Mozhi\Test;
 
 use Aheenam\Mozhi\RouteResolver;
 use Illuminate\Support\Facades\Storage;
-use Aheenam\Mozhi\Documents\MarkdownDocument;
+use Aheenam\Mozhi\Documents\MarkdownDocument\MarkdownDocument;
 use Illuminate\Contracts\Filesystem\Filesystem;
 
 class RouteResolverTest extends TestCase
